refactor: modularize codebase and improve UX
